Allow MPTCP without timestamps
In fact, the receiver is able to estimate the RTT without timestamps.
Because, we need rcv_rtt_est.rtt only to calculate the receive-window.
As a full receive-window is sent in the timeperiod of a maximum RTT (max
among the subflows), the TCP rcv-rtt estimator estimates exactly this
maximum RTT. Which is basically what we need.

Of course, using timestamps is still much more accurate.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>