Fix: Don't send reinjected segments that already reached the peer
They may end up in the reinject-queue, if the RTO happens after a
meta-reinjection or rbuf-opti-reinjection. In that case we should not
try to send the segment, because it may block sending. The segment
coming from the reinject queue has already been queued on all subflows
and thus get_available_subflow won't find a segment and we are stuck
with this segment in the reinject-queue.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>