Fix: Set path_mask to 0, before pushing to subflow
Since 6435ad76c54eb9, the path-mask is a union with inet_skb_param. As
the subskb is a clone of the meta's skb, the path_mask may be != 0 and
thus inet_skb_param is no more initialized to 0 as it should.

We fix this, by resetting the path-mask to 0 just before pushing the
segment on the subflow in mptcp_write_xmit.

Special attention must be paid to reinjected segments.
They are coming from the subflow (thus path_mask == 0), however we don't
want them to be reinjected twice on the same subflow. Thus, before
continuing with reinjection in mptcp_write_xmit, we get the path_mask
out of the meta-write-queue and set it on the reinject-segment.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Cc: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>