Fix: Always lock the socket in mptcp_inherit_sk
Previously we did not lock if it was the meta-sk on the server-side.

This creates the following race:
In mptcp_check_req_master we allocate the mpcb and set child_tp->mpc to
1. Thus, if a new segment comes in while still being inside
mptcp_check_req_master, this segment will take the meta-lock in
tcp_v4_rcv. So, we are processing the third ack and this new segment in
the master-sk at the same time.

Fix is easy:
Always take the lock in mptcp_inherit_sk and unlock it again in
tcp_child_process.

There is one problem with lockdep-checks. As we take the meta-lock we
take three locks of the same class: The listener-sock in tcp_v4_rcv, the
master-sk in sk_clone and finally the meta-sk in mptcp_inherit_sk. Thus,
lockdep will complain.

We create a custom lock-class for the meta-sk. That way lockdep won't
complain anymore.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Reviewed-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>