Fix cnt_established accounting
The problem is, that if we enter mptcp_set_state and sk->sk_state is
already TCP_CLOSE, and we somehow try again to set the state to closed,
cnt_established will get decremented twice for this subsocket.

This fix contains multiple changes:

1. Check for tp->mpc at the caller of mptcp_set_state and not inside
   mptcp_set_state itself.
2. In mptcp_sub_close, prevent calling tcp_close_state, if the state of
   the subsocket is already TCP_CLOSE. It's pointless then.
3. In mptcp_set_state, don't change cnt_established, if sk->sk_state is
   already TCP_CLOSE. Because, if it is already closed, mptcp_set_state
   has already decremented cnt_estblished.

Additionally, I use tp instead of tcp_sk(sk) in mptcp_set_state.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>