Fix: Handle sock_owned_by_user in mptcp_pm_addr_event_handler
If the socket is owned by the user and we enter
mptcp_pm_addr_event_handler, the sockets will get corrupted as we are
inside the critical region from the bottom-half address-handler and the
user-context.

We fix this, by adding the usual check of sock_owned_by_user in
mptcp_pm_addr_event_handler.
If the socket is owned by the user, we create a work-queue that will
handle the address-update.

However, we cannot pass the ptr from the addr-handler to the work-queue,
as he may have disappeared when the work-queue will get executed. Or a
new address-event may have happened and for example the address has come
up again.

So, in the work-queue we iterate over all addresses and check if they
exist inside the mpcb->addr4. If not, we add it, or update it.
Then we iterate over the addresses in mpcb->addr4 and check if there
exists a corresponding address on one of the interfaces. If not, we
remove this address.

This patch introduces quite some code-duplication in the
address-handling code. I will refactor later.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>