Fix Bug#81: Add token in hash-table in tcp_connect_init
And change the locking of the token-table.
We had two locks. One for the request-sock token-table and one for the
meta-socket token-table.
We always took both locks in the most speed-critical regions
(tcp_openreq_init and tcp_connect_init). Thus, its better to protect
both token-tables by a single lock.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>