Use sk_data_ready-callback
This simplifies the receive-path significantly inside tcp_data_queue and
we thus have a much cleaner layer separation.

In tcp_fin we have to do some changes, because a data-fin might be
combined with a subflow-fin. This subflow-fin might close the subflow.
We have to prevent this, to allow sending a data-ack of the data-fin.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>